updated file structures and added backend deployment links
